+++ systemtest/src/main/java/io/strimzi/systemtest/BaseST.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test;
+package io.strimzi.systemtest;
 
 import io.strimzi.test.k8s.KubeClusterResource;
 import org.junit.jupiter.api.BeforeAll;
